Moved user create form to separate component Auto log out user when registering
